Use Int in EventHandler ids & change order of api disabling to call children api first

+++ Common/src/main/java/com/github/serivesmejia/eocvsim/util/event/EventListener.kt
@@ -24,7 +24,7 @@
 package com.github.serivesmejia.eocvsim.util.event
 
 @JvmInline
-value class EventListenerId(val value: Long)
+value class EventListenerId(val value: Int)
 
 typealias OnceEventListener = () -> Unit
 typealias EventListener = EventListenerContext.() -> Unit
@@ -34,7 +34,6 @@ typealias EventListener = EventListenerContext.() -> Unit
  * to allow removing itself from the event handler
  * @param handler the event handler
  * @param id the listener ID
- * @param isOnceListener whether the listener is a once listener
  */
 class EventListenerContext(
     private val handler: EventHandler,
